% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_functions.R
\docType{package}
\name{bomrang-deprecated}
\alias{bomrang-deprecated}
\alias{bomrang_cache_list}
\alias{bomrang-deprecated-package}
\alias{bomrang-deprecated}
\alias{bomrang_cache_details}
\alias{bomrang-deprecated}
\alias{bomrang_cache_delete}
\alias{bomrang-deprecated}
\alias{bomrang_cache_delete_all}
\title{Deprecated function(s) in the bomrang package}
\usage{
bomrang_cache_list()

bomrang_cache_details()

bomrang_cache_delete()

bomrang_cache_delete_all()
}
\description{
These functions are now deprecated in \pkg{bomrang}.
}
\section{Details}{

\tabular{rl}{
  \code{bomrang_cache_list} \tab now superceded by \code{manage_cache$list}\cr
  \code{bomrang_cache_details} \tab now superceded by \code{manage_cache$details}\cr
  \code{bomrang_cache_delete} \tab now superceded by \code{manage_cache$delete}\cr
  \code{bomrang_cache_delete_all} \tab now superceded by \code{manage_cache$delete_all }\cr
}
}

