% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bomrang-package.R
\docType{package}
\name{bomrang-package}
\alias{bomrang}
\alias{bomrang-package}
\title{bomrang: Australian Government Bureau of Meteorology ('BOM') Data Client}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

Provides functions to interface with Australian Government Bureau
    of Meteorology ('BOM') data, fetching data and returning a data frame of
    precis forecasts, historical and current weather data from stations,
    agriculture bulletin data, 'BOM' 0900 or 1500 weather bulletins and
    downloading and importing radar and satellite imagery files.  Data (c)
    Australian Government Bureau of Meteorology Creative Commons (CC)
    Attribution 3.0 licence or Public Access Licence (PAL) as appropriate.  See
    <http://www.bom.gov.au/other/copyright.shtml> for further details.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ropensci/bomrang}
  \item \url{https://docs.ropensci.org/bomrang/}
  \item Report bugs at \url{https://github.com/ropensci/bomrang/issues}
}

}
\author{
\strong{Maintainer}: Adam H. Sparks \email{adamhsparks@gmail.com} (\href{https://orcid.org/0000-0002-0061-8359}{ORCID})

Authors:
\itemize{
  \item Jonathan Carroll \email{rpkg@jcarroll.com.au} (\href{https://orcid.org/0000-0002-1404-5264}{ORCID})
  \item James Goldie \email{me@rensa.co} (\href{https://orcid.org/0000-0002-5024-6207}{ORCID})
  \item Dean Marchiori \email{deanmarchiori@gmail.com}
  \item Paul Melloy \email{paul@melloy.com.au} (\href{https://orcid.org/0000-0003-4253-7167}{ORCID})
  \item Mark Padgham \email{mark.padgham@email.com} (\href{https://orcid.org/0000-0003-2172-5265}{ORCID}) [reviewer]
  \item Hugh Parsonage \email{hugh.parsonage@gmail.com} (\href{https://orcid.org/0000-0003-4055-0835}{ORCID})
  \item Keith Pembleton \email{keith.pembleton@usq.edu.au} (\href{https://orcid.org/0000-0002-1896-4516}{ORCID})
}

Other contributors:
\itemize{
  \item James Balamuta \email{balamut2@illinois.edu} (\href{https://orcid.org/0000-0003-2826-8458}{ORCID}) [contributor]
  \item Brooke Anderson \email{brooke.anderson@colostate.edu} (\href{https://orcid.org/0000-0002-5012-9035}{ORCID}) [reviewer]
}

}
\keyword{internal}
