% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BOMRang-package.r
\docType{package}
\name{BOMRang}
\alias{BOMRang}
\alias{BOMRang-package}
\title{Fetches Australian Government Bureau of Meteorology (BOM) XML précis forecast
and returns a tidy data frame of the daily weather forecast.}
\description{
Fetches Australian Government Bureau of Meteorology (BOM) XML précis forecast
and returns a tidy data frame of the daily weather forecast.
}
\seealso{
\strong{Useful links:}
\itemize{
\item{\url{https://github.com/ToowoombaTrio/BOMRang}}
\item{Report bugs at \url{https://github.com/ToowoombaTrio/BOMRang/issues}}
}
}
\author{
Adam H Sparks and Keith Pembleton
}
