% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr.R
\docType{import}
\name{filter}
\alias{filter}
\alias{filter.bomrang_tbl}
\title{Objects exported from other packages}
\usage{
\method{filter}{bomrang_tbl}(.data, ...)
}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{dplyr}{\code{\link[dplyr]{filter}}}
}}

