% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ag_bulletin.R
\name{get_ag_bulletin}
\alias{get_ag_bulletin}
\title{Get BOM agriculture bulletin information for select stations}
\usage{
get_ag_bulletin(state = "AUS", filepath = NULL)
}
\arguments{
\item{state}{Australian state or territory as full name or postal code.
Fuzzy string matching via \code{\link[base]{agrep}} is done.  Defaults to
"AUS" returning all state bulletins, see Details for more.}

\item{filepath}{A character string of the location of \acronym{XML} file(s)
to parse.  If \var{filepath} is specified function will use \acronym{BOM}
daily précis forecast from a local \acronym{XML} file at the specified
location and not the \acronym{BOM} \acronym{FTP} server. See Details for
more.}
}
\value{
Tidy \code{\link[data.table]{data.table}} of Australia \acronym{BOM} 
 agricultural bulletin information.  For full details of fields and units
 returned see Appendix 3 in the \pkg{bomrang} vignette, use \cr
 \code{vignette("bomrang", package = "bomrang")} to view.
}
\description{
Fetch the \acronym{BOM} agricultural bulletin information and return it in a
data frame
}
\details{
Allowed state and territory postal codes, only one state per request
or all using \code{AUS}.
 \describe{
   \item{ACT}{Australian Capital Territory (will return NSW)}
   \item{NSW}{New South Wales}
   \item{NT}{Northern Territory}
   \item{QLD}{Queensland}
   \item{SA}{South Australia}
   \item{TAS}{Tasmania}
   \item{VIC}{Victoria}
   \item{WA}{Western Australia}
   \item{AUS}{Australia, returns forecast for all states, NT and ACT}
 }

 In some situations, access may be restricted to insecure \acronym{FTP}
 connections. Using \var{filepath} allows you to download and save the
 \acronym{XML} files locally for use in \pkg{bomrang}.
}
\examples{
\donttest{
ag_bulletin <- get_ag_bulletin(state = "QLD")
ag_bulletin
}

}
\references{
Agricultural observations are retrieved from the Australian Bureau of
Meteorology (\acronym{BOM}) Weather Data Services Agriculture Bulletins, \cr
\url{http://www.bom.gov.au/catalogue/observations/about-agricultural.shtml}

and

Australian Bureau of Meteorology (\acronym{BOM})) Weather Data Services
Observation of Rainfall, \cr
\url{http://www.bom.gov.au/climate/how/observations/rain-measure.shtml}

Station location and other metadata are sourced from the Australian Bureau of
Meteorology (\acronym{BOM}) webpage, Bureau of Meteorology Site Numbers: \cr
\url{http://www.bom.gov.au/climate/cdo/about/site-num.shtml}
}
\author{
Adam H. Sparks, \email{adamhsparks@gmail.com} and Paul Melloy
\email{paul.melloy@usq.edu.au}
}
