% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_satellite_imagery.R
\name{get_available_images}
\alias{get_available_images}
\title{Get a listing of Available BoM Satellite Imagery}
\usage{
get_available_images(product_id = NULL)
}
\arguments{
\item{product_id}{Character.  BoM product ID to download in GeoTIFF format
and import as a \code{\link{raster}} object, either as a single
\code{\link{raster::raster}} layer or \code{\link{raster::stack}} object.
Defaults to all images available for a requested BoM Product ID.}
}
\value{
A vector of all available files for the requested Product ID.
}
\description{
Fetch a listing of BoM satellite imagery from
\url{ftp://ftp.bom.gov.au/anon/gen/gms/} to determine which files are
currently available for download.  Files are available at ten minute update
frequency with a 24 hour delete time.  Useful to know the most recent files
available and then specify in the \code{\link{get_satellite_imagery}}
function.
}
\details{
Valid BoM satellite Product IDs include:
\describe{
\item{IDE00420}{AHI cloud cover only 2km FD GEOS GIS}
\item{IDE00421}{AHI IR (Ch13) greyscale 2km FD GEOS GIS}
\item{IDE00422}{AHI VIS (Ch3) greyscale 2km FD GEOS GIS}
\item{IDE00423}{AHI IR (Ch13) Zehr 2km FD GEOS GIS}
\item{IDE00425}{AHI VIS (true colour) / IR (Ch13 greyscale) composite 1km FD
GEOS GIS}
\item{IDE00426}{AHI VIS (true colour) / IR (Ch13 greyscale) composite 2km FD
GEOS GIS}
\item{IDE00427}{AHI WV (Ch8) 2km FD GEOS GIS}
\item{IDE00430}{AHI cloud cover only 2km AUS equirect. GIS}
\item{IDE00431}{AHI IR (Ch13) greyscale 2km AUS equirect. GIS}
\item{IDE00432}{AHI VIS (Ch3) greyscale 2km AUS equirect. GIS}
\item{IDE00433}{AHI IR (Ch13) Zehr 2km AUS equirect. GIS}
\item{IDE00435}{AHI VIS (true colour) / IR (Ch13 greyscale) composite 1km AUS
equirect. GIS}
\item{IDE00436}{AHI VIS (true colour) / IR (Ch13 greyscale) composite 2km AUS
equirect. GIS}
\item{IDE00437}{AHI WV (Ch8) 2km AUS equirect. GIS}
\item{IDE00439}{AHI VIS (Ch3) greyscale 0.5km AUS equirect. GIS}
}
}
\examples{
\dontrun{
Check availability of AHI VIS (true colour) / IR (Ch13 greyscale) composite
1km FD GEOS GIS images
imagery <- available_images(product_id = "IDE00425")
}

}
\references{
Himawari-8 and -9 Facts and Figures,
http://www.bom.gov.au/australia/satellite/himawari.shtml
}
\author{
Adam H Sparks, \email{adamhsparks@gmail.com}
}
