% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_BOM_forecast.R
\name{get_BOM_QLD}
\alias{get_BOM_QLD}
\title{Get BOM Forecast for Queensland}
\usage{
get_BOM_QLD()
}
\value{
Data frame of a Australia BOM forecast for Queensland for max temperature,
min temperature and corresponding locations with lat/lon values for the next
six days.
}
\description{
Fetch the BOM forecast and create a data frame object that can be used for
interpolating.
}
\examples{
\dontrun{
BOM_forecast <- get_BOM_QLD()
}
}
