% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bulletin.R
\name{get_bulletin}
\alias{get_bulletin}
\title{Get BOM Agriculture Bulletin}
\usage{
get_bulletin(state = NULL)
}
\arguments{
\item{state}{Australian state or territory as postal code, see details for
instruction.}
}
\value{
Data frame of a Australia BOM agricultural bulletin information

\describe{
   # to be updated
}
}
\description{
Fetch the BOM agricultural bulletin information
}
\details{
Allowed state and territory postal codes, only one state per request
or all using \code{AUS}.
 \describe{
   \item{NSW}{New South Wales}
   \item{NT}{Northern Territory}
   \item{QLD}{Queensland}
   \item{SA}{South Australia}
   \item{TAS}{Tasmania}
   \item{VIC}{Victoria}
   \item{WA}{Western Australia}
   \item{AUS}{Australia, returns forecast for all states}
 }
}
\examples{
\dontrun{
ag_bulletin <- get_bulletin(state = "QLD")
}

}
\references{
Australian Bureau of Meteorology (BOM) Weather Data Services
\url{http://www.bom.gov.au/catalogue/data-feeds.shtml}
}
\author{
Adam H Sparks, \email{adamhsparks@gmail.com} and Keith Pembleton \email{keith.pembleton@usq.edu.au}
}
