% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_coastal_forecast.R
\name{get_coastal_forecast}
\alias{get_coastal_forecast}
\title{Get BOM coastal waters forecast}
\usage{
get_coastal_forecast(state = "AUS", filepath = NULL)
}
\arguments{
\item{state}{Australian state or territory as full name or postal code.
Fuzzy string matching via \code{\link[base]{agrep}} is done.  Defaults to
"AUS" returning all state forecasts, see details for further information.}

\item{filepath}{A character string of the location of \acronym{XML} file(s)
to parse.  If \var{filepath} is specified function will use \acronym{BOM}
daily précis forecast from a local \acronym{XML} file at the specified
location and not the \acronym{BOM} \acronym{FTP} server. See Details for
more.}
}
\value{
Tidy \code{\link[data.table]{data.table}} of a Australia \acronym{BOM}
Coastal Waters Forecast. For full details of fields and units
returned see Appendix 5 in the \pkg{bomrang} vignette, use \cr
\code{vignette("bomrang", package = "bomrang")} to view.
}
\description{
Fetch the \acronym{BOM} daily Coastal Waters Forecast and return a data frame
of the forecast regions for a specified state or region.
}
\details{
Allowed state and territory postal codes, only one state per request
or all using \code{AUS}.
 \describe{
   \item{ACT}{Australian Capital Territory (will return NSW)}
   \item{NSW}{New South Wales}
   \item{NT}{Northern Territory}
   \item{QLD}{Queensland}
   \item{SA}{South Australia}
   \item{TAS}{Tasmania}
   \item{VIC}{Victoria}
   \item{WA}{Western Australia}
   \item{AUS}{Australia, returns forecast for all states, NT and ACT}
 }

 In some situations, access may be restricted to insecure \acronym{FTP}
 connections. Using \var{filepath} allows you to download and save the
 \acronym{XML} files locally for use in \pkg{bomrang}.
}
\examples{
\donttest{
coastal_forecast <- get_coastal_forecast(state = "NSW")
coastal_forecast
}
}
\references{
Forecast data come from Australian Bureau of Meteorology (BOM) Weather Data
Services \cr
\url{http://www.bom.gov.au/catalogue/data-feeds.shtml}

Location data and other metadata come from the \acronym{BOM} anonymous
\acronym{FTP} server with spatial data \cr
\url{ftp://ftp.bom.gov.au/anon/home/adfd/spatial/}, specifically the
\acronym{DBF} file portion of a shapefile, \cr
\url{ftp://ftp.bom.gov.au/anon/home/adfd/spatial/IDM00003.dbf}
}
\author{
Dean Marchiori, \email{deanmarchiori@gmail.com} and Paul Melloy
\email{paul@melloy.com.au}
}
