% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_radar_imagery.R
\name{get_radar_imagery}
\alias{get_radar_imagery}
\title{Get BOM Radar Imagery}
\usage{
get_radar_imagery(product_id = NULL)
}
\arguments{
\item{product_id}{Character.  BOM product ID to download and import as a 
\code{\link[raster]{raster}} object. Value is required.}
}
\value{
A raster layer based on the most recent `.gif` radar image snapshot published
by the BOM.
}
\description{
Fetch BOM radar imagery from
\url{ftp://ftp.bom.gov.au/anon/gen/radar/} and return a raster
\code{\link[raster]{raster}} object. Files available are the most recent 
radar snapshot which are updated approximately every 6 to 10 minutes. 
Suggested to check file availability first by using
\code{\link{get_available_radar}}.
}
\details{
Valid BOM satellite Product IDs for radar imagery can be obtained
from \code{\link{get_available_radar}}.
}
\examples{
\dontrun{
# Fetch most recent radar image for Wollongong 256km radar

imagery <- get_radar_imagery(product_id = "IDR032")  
raster::plot(imagery)

}

}
\references{
Australian Bureau of Meteorology (BOM) radar images
\url{http://www.bom.gov.au/australia/radar/}
}
\seealso{
\code{\link{get_available_radar}}
}
\author{
Dean Marchiori, \email{deanmarchiori@gmail.com}
}
