% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr.R
\docType{import}
\name{group_by}
\alias{group_by}
\alias{group_by.bomrang_tbl}
\title{Objects exported from other packages}
\usage{
\method{group_by}{bomrang_tbl}(.data, ...)
}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{dplyr}{\code{\link[dplyr]{group_by}}}
}}

