% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{JSONurl_latlon_by_station_name}
\alias{JSONurl_latlon_by_station_name}
\title{JSONurl_latlon_by_station_name}
\format{An object of class \code{data.table} (inherits from \code{data.frame}) with 886 rows and 14 columns.}
\source{
\url{ftp://ftp.bom.gov.au/anon2/home/ncc/metadata/sitelists/stations.zip}
}
\usage{
JSONurl_latlon_by_station_name
}
\description{
A database of URLs that provide a JSON file of current weather conditions
for BoM weather stations.
}
\details{
\describe{
   \item{site}{Unique BoM identifier for each station}
   \item{dist}{BoM rainfall district}
   \item{name}{BoM station name}
   \item{start}{Year data collection starts}
   \item{end}{Year data collection ends (will always be current)}
   \item{state}{State name (postal code abbreviation)}
   \item{lat}{Latitude (decimal degrees)}
   \item{lon}{Longitude (decimal degrees)}
   \item{elev_m}{Station elevation (metres)}
   \item{bar_ht}{Bar height (metres)}
   \item{wmo}{World Meteorological Organization number (unique ID used worldwide)}
   \item{state_code}{BoM code used to identify states and territories}
   \item{url}{URL that serves JSON file of station weather data}
   }
}
\keyword{datasets}
