% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr.R
\docType{import}
\name{mutate}
\alias{mutate}
\alias{mutate.bomrang_tbl}
\title{Objects exported from other packages}
\usage{
\method{mutate}{bomrang_tbl}(.data, ...)
}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{dplyr}{\code{\link[dplyr]{mutate}}}
}}

