% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr.R
\docType{import}
\name{slice}
\alias{slice}
\alias{slice.bomrang_tbl}
\title{Objects exported from other packages}
\usage{
\method{slice}{bomrang_tbl}(.data, ...)
}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{dplyr}{\code{\link[dplyr]{slice}}}
}}

