% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sweep_for_forecast_towns.R
\name{sweep_for_forecast_towns}
\alias{sweep_for_forecast_towns}
\title{Find nearest BOM forecast towns}
\usage{
sweep_for_forecast_towns(latlon = c(-35.3, 149.2))
}
\arguments{
\item{latlon}{A length-2 numeric vector. By default, Canberra
(approximately).}
}
\value{
A \code{\link{data.table}} of all forecast towns (in this package)
sorted by distance from \var{latlon}, ascending.
}
\description{
Find nearest BOM forecast towns
}
\author{
Hugh Parsonage, \email{hugh.parsonage@gmail.com} and
 James Goldie, \email{me@rensa.co}
}
