% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_forecast_locations.R
\name{update_forecast_locations}
\alias{update_forecast_locations}
\title{Update internal database with latest BoM forecast locations}
\usage{
update_forecast_locations()
}
\value{
Updated internal database of BoM précis forecast locations
}
\description{
Download the latest forecast locations from the BoM server and update
bomrang's internal database of précis forecast locations,
\code{\link{AAC_codes}} used by \code{\link{get_precis_forecast}}.  There is
no need to use this unless you know that a forecast location exists in a
more current version of the BoM précis forecast location database that is not
available in the database distributed with \code{\link{bomrang}}.
}
\examples{
\dontrun{
update_forecast_locations()
}
}
\references{
Australian Bureau of Meteorology (BoM) Weather Data Services
\url{http://www.bom.gov.au/catalogue/data-feeds.shtml}
}
\author{
Adam H Sparks, \email{adamhsparks@gmail.com}
}
