% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_locations.R
\name{update_locations}
\alias{update_locations}
\title{Update Internal Database With Latest BOM Forecast Locations From BOM FTP Server}
\usage{
update_locations()
}
\value{
Updated internal database of BOM forecast locations
}
\description{
This function downloads the latest forecast locations from the BOM server
and updates BOMRang's internal database of forecast locations.  There is no
need to use this unless you know that a forecast location exists in a more
current version of the BOM forecast location database that is not available
in the database distributed with \code{\link{BOMRang}}.
}
\examples{
\dontrun{
update_locations()
}
}
\references{
Australian Bureau of Meteorology (BOM) Weather Data Services
\url{http://www.bom.gov.au/catalogue/data-feeds.shtml}
}
\author{
Adam H Sparks, \email{adamhsparks@gmail.com}
}
