% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_precis_locations.R
\name{update_precis_locations}
\alias{update_precis_locations}
\title{Update Internal Database With Latest BOM Forecast Locations From BOM FTP Server}
\usage{
update_precis_locations()
}
\value{
Updated internal database of BOM précis forecast locations
}
\description{
Download the latest forecast locations from the BOM server and update
bomrang's internal database of précis forecast locations.  There is no
need to use this unless you know that a forecast location exists in a more
current version of the BOM précis forecast location database that is not available
in the database distributed with \code{\link{bomrang}}.
}
\examples{
\dontrun{
update_precis_locations()
}
}
\references{
Australian Bureau of Meteorology (BOM) Weather Data Services
\url{http://www.bom.gov.au/catalogue/data-feeds.shtml}
}
\author{
Adam H Sparks, \email{adamhsparks@gmail.com}
}
