% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_station_locations.R
\name{update_station_locations}
\alias{update_station_locations}
\title{Update bomrang Internal Databases with Latest BoM Station Metadata}
\usage{
update_station_locations()
}
\value{
Updated internal databases of BoM station locations and JSON URLs
}
\description{
Download the latest station locations and metadata and update bomrang's
internal databases that support the use of \code{\link{get_current_weather}}
and \code{\link{get_ag_bulletin}}.  There is no need to use this unless you
know that a station exists in BoM's database that is not available in the
databases distributed with \code{\link{bomrang}}.
}
\details{
If \code{\link[sf]{sf}} is installed locally, this function will
automatically check and correct any invalid state values for stations located
in Australia. If \code{\link[sf]{sf}} is not installed, the function will
update the internal database without validating the state values for stations
by reported lon/lat location.
}
\examples{
\dontrun{
update_station_locations()
}
}
\references{
Station location and other metadata are sourced from the Australian Bureau of
Meteorology (BoM) webpage, Bureau of Meteorology Site Numbers:
\url{http://www.bom.gov.au/climate/cdo/about/site-num.shtml}
}
\author{
Adam H Sparks, \email{adamhsparks@gmail.com}
}
