% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{AAC_codes}
\alias{AAC_codes}
\title{AAC_codes}
\format{A data frame with 1369 observations of 4 elements:
\describe{
  \item{AAC}{Unique identifier for each location}
  \item{PT_NAME}{Human readable location name}
  \item{LON}{Longitude}
  \item{LAT}{Latitude}
}

The \code{AAC_codes} data are automatically loaded with the
\code{\link{BOMRang}} package and merged with the latest available forecast
from the BOM.}
\source{
\url{ftp://ftp.bom.gov.au/anon/home/adfd/spatial/IDM00013.dbf}
}
\usage{
AAC_codes
}
\description{
AAC_codes
}
\keyword{datasets}
