% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bomrang.r
\docType{package}
\name{bomrang-package}
\alias{bomrang}
\alias{bomrang-package}
\title{bomrang: Australian Government Bureau of Meteorology (BOM) Data from R}
\description{
Provides functions to interface with Australian Government Bureau
of Meteorology (BOM) data, fetching data and returning a tidy data frame of
précis forecasts, historical and current weather data from stations,
agriculture bulletin data, BOM 0900 or 1500 weather bulletins and
downloading and importing radar and satellite imagery files. Data (c)
Australian Government Bureau of Meteorology Creative Commons (CC)
Attribution 3.0 licence or Public Access Licence (PAL) as appropriate. See
<http://www.bom.gov.au/other/copyright.shtml> for further details.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ropensci/bomrang}
  \item \url{https://ropensci.github.io/bomrang/}
  \item Report bugs at \url{https://github.com/ropensci/bomrang/issues}
}

}
\author{
\strong{Maintainer}: Adam Sparks \email{adamhsparks@gmail.com} (0000-0002-0061-8359)

Authors:
\itemize{
  \item Jonathan Carroll \email{rpkg@jcarroll.com.au} (0000-0002-1404-5264)
  \item Dean Marchiori \email{deanmarchiori@gmail.com}
  \item Mark Padgham \email{mark.padgham@email.com} (0000-0003-2172-5265) [NA]
  \item Hugh Parsonage \email{hugh.parsonage@gmail.com} (0000-0003-4055-0835)
  \item Keith Pembleton \email{keith.pembleton@usq.edu.au} (0000-0002-1896-4516)
}

Other contributors:
\itemize{
  \item James Balamuta \email{balamut2@illinois.edu} (0000-0003-2826-8458) [contributor]
  \item Brooke Anderson \email{brooke.anderson@colostate.edu} (0000-0002-5012-9035) [NA]
}

}
