% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BOMRang-package.r
\docType{package}
\name{BOMRang}
\alias{BOMRang}
\alias{BOMRang-package}
\title{Fetches Fetch Australian Government Bureau of Meteorology Weather
data, currently Queensland only forecast and returns a tidy data frame of
next seven days weather forecast.}
\description{
Fetches Fetch Australian Government Bureau of Meteorology Weather
data, currently Queensland only forecast and returns a tidy data frame of
next seven days weather forecast.
}
\seealso{
\strong{Useful links:}
\itemize{
\item{\url{https://github.com/ToowoombaTrio/BOMRang}}
\item{Report bugs at \url{https://github.com/ToowoombaTrio/BOMRang/issues}}
}
}
\author{
Adam Sparks and Keith Pembleton
}
