% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bomrang.r
\docType{package}
\name{bomrang}
\alias{bomrang}
\alias{bomrang-package}
\title{Australian Government Bureau of Meteorology (BOM) Data from R}
\description{
Australian Government Bureau of Meteorology (BOM) Data from R
}
\author{
Adam H Sparks and Jonathan Carroll and  Mark Padgham and Hugh
Parsonage and Keith Pembleton
}
\seealso{
\strong{Useful links:}
\itemize{
\item{Development repository: \url{https://github.com/ropensci/bomrang}}
\item{Static documentation: \url{https://ropensci.github.io/bomrang/}}
\item{Report bugs at \url{https://github.com/ropensci/bomrang/issues}}
}
}

