% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bomrang.r
\docType{package}
\name{bomrang}
\alias{bomrang}
\alias{bomrang-package}
\title{Fetches Australian Government Bureau of Meteorology (BoM) Weather Data Files
and Returns a Tidy Data Frame or Raster Stack Object of the Data.}
\description{
Fetches Australian Government Bureau of Meteorology (BoM) Weather Data Files
and Returns a Tidy Data Frame or Raster Stack Object of the Data.
}
\seealso{
\strong{Useful links:}
\itemize{
\item{\url{https://github.com/ropensci/bomrang}}
\item{Report bugs at \url{https://github.com/ropensci/bomrang/issues}}
}
}
\author{
Adam H Sparks and Hugh Parsonage and Keith Pembleton
}
