% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bomrang-package.r
\docType{package}
\name{bomrang}
\alias{bomrang}
\alias{bomrang-package}
\title{Fetches Australian Government Bureau of Meteorology (BOM) XML weather data
files and returns a tidy data frame of the data.}
\description{
Fetches Australian Government Bureau of Meteorology (BOM) XML weather data
files and returns a tidy data frame of the data.
}
\seealso{
\strong{Useful links:}
\itemize{
\item{\url{https://github.com/ToowoombaTrio/bomrang}}
\item{Report bugs at \url{https://github.com/ToowoombaTrio/bomrang/issues}}
}
}
\author{
Adam H Sparks and Keith Pembleton
}
