% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bomrang.r
\docType{package}
\name{bomrang}
\alias{bomrang}
\alias{bomrang-package}
\title{Australian government Bureau of Meteorology (\acronym{BOM}) data client}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

Provides functions to interface with Australian Government Bureau
    of Meteorology ('BOM') data, fetching data and returning a tidy data frame
    of precis forecasts, historical and current weather data from stations,
    agriculture bulletin data, 'BOM' 0900 or 1500 weather bulletins and
    downloading and importing radar and satellite imagery files.  Data (c)
    Australian Government Bureau of Meteorology Creative Commons (CC)
    Attribution 3.0 licence or Public Access Licence (PAL) as appropriate.  See
    <http://www.bom.gov.au/other/copyright.shtml> for further details.
}
\seealso{
\strong{Useful links:}
\itemize{
\item{Development repository: \url{https://github.com/ropensci/bomrang}}
\item{Static documentation: \url{https://docs.ropensci.org/bomrang/}}
\item{Report bugs: \url{https://github.com/ropensci/bomrang/issues}}
}
}
\author{
Adam H. Sparks and Jonathan Carroll and James Goldie and Dean
Marchiori and Paul Melloy and Mark Padgham and Hugh Parsonage and Keith
Pembleton
}
