% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/current_weather.R
\name{current_weather}
\alias{current_weather}
\title{Current weather observations of a station}
\usage{
current_weather(station_name, raw = FALSE)
}
\arguments{
\item{station_name}{The name of the weather station. Fuzzy string matching via
\code{base::agrep} is done.}

\item{raw}{Do not convert the columns \code{data.table} to the appropriate classes. (\code{FALSE} by default.)}
}
\description{
Current weather observations of a station
}
\details{
Note that the column \code{local_date_time_full} is set to a \code{POSIXct} object
in the local time of the \strong{user}.
}
\examples{
\dontrun{
  # warning
  Melbourne_forecast <- current_weather("Melbourne")

  # no warning
  Melbourne_forecast <- current_weather("Melbourne (Olympic Park)")
}
}
