% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ag_bulletin.R
\name{get_ag_bulletin}
\alias{get_ag_bulletin}
\title{Get BOM Agriculture Bulletin}
\usage{
get_ag_bulletin(state = NULL)
}
\arguments{
\item{state}{Australian state or territory as postal code, see details for
instruction.}
}
\value{
Data frame of a Australia BOM agricultural bulletin information

\describe{
   \item{obs-time-local}{Observation time}
   \item{obs-time-utc}{Observation time (time in UTC)}
   \item{time-zone}{Time zone for observation}
   \item{site}{Unique BOM identifier for each station}
   \item{dist}{BOM rainfall district}
   \item{station}{BOM station name}
   \item{start}{Year data collection starts}
   \item{end}{Year data collection ends (will always be current)}
   \item{state}{State name (postal code abbreviation)}
   \item{lat}{Latitude (decimal degrees)}
   \item{lon}{Longitude (decimal degrees)}
   \item{elev_m}{Station elevation (metres)}
   \item{bar_ht}{Bar height (metres)}
   \item{WMO}{World Meteorological Organization number (unique ID used worldwide)}
   \item{r}{Rain to 9am (millimetres). \strong{Trace will be reported as 0.01}}
   \item{tn}{Minimum temperature (degrees Celsius)}
   \item{tx}{Maximum temperature (degrees Celsius)}
   \item{twd}{Wet bulb depression (degrees Celsius)}
   \item{ev}{Evaporation (millimetres)}
   \item{tg}{Terrestrial minimum temperature (degrees Celsius)}
   \item{sn}{Sunshine (hours)}
   \item{t5}{5cm soil temperature (degrees Celsius)}
   \item{t10}{10cm soil temperature (degrees Celsius)}
   \item{t20}{20cm soil temperature (degrees Celsius)}
   \item{t50}{50cm soil temperature (degrees Celsius)}
   \item{t1m}{1m soil temperature (degrees Celsius)}
   \item{wr}{Wind run (kilometres)}
}
}
\description{
Fetch the BOM agricultural bulletin information and return a tidy data frame
}
\details{
Allowed state and territory postal codes, only one state per request
or all using \code{AUS}.
 \describe{
   \item{NSW}{New South Wales}
   \item{NT}{Northern Territory}
   \item{QLD}{Queensland}
   \item{SA}{South Australia}
   \item{TAS}{Tasmania}
   \item{VIC}{Victoria}
   \item{WA}{Western Australia}
   \item{AUS}{Australia, returns bulletin for all states}
 }
}
\examples{
\dontrun{
ag_bulletin <- get_ag_bulletin(state = "QLD")
}

}
\references{
Australian Bureau of Meteorology (BOM) Weather Data Services Agriculture Bulletins
\url{http://www.bom.gov.au/catalogue/observations/about-agricultural.shtml}

Australian Bureau of Meteorology (BOM) Weather Data Services Observation of Rainfall
\url{http://www.bom.gov.au/climate/how/observations/rain-measure.shtml}
}
\author{
Adam H Sparks, \email{adamhsparks@gmail.com}
}
