% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bulletin.R
\name{get_bulletin}
\alias{get_bulletin}
\title{Get BOM Agriculture Bulletin}
\usage{
get_bulletin(state = NULL)
}
\arguments{
\item{state}{Australian state or territory as postal code, see details for
instruction.}
}
\value{
Data frame of a Australia BOM agricultural bulletin information

\describe{
   \item{obs-time-utc}{Observation time (Time in UTC)}
   \item{time-zone}{Time zone for observation}
   \item{site}{Unique BOM identifier for each station}
   \item{name}{BOM station name}
   \item{r}{Rain to 9am (millimetres). \strong{Trace will be reported as 0.01}}
   \item{tn}{Minimum temperature (Celsius)}
   \item{tx}{Maximum temperature (Celsius)}
   \item{twd}{Wetbulb depression (Celsius)}
   \item{ev}{Evaporation (millimetres)}
   \item{tg}{Terrestrial minimum temperature (Celsius)}
   \item{sn}{Sunshine (Hours)}
   \item{t5}{5cm temperature (Celsius)}
   \item{t10}{10cm temperature (Celsius)}
   \item{t20}{20cm temperature (Celsius)}
   \item{t50}{50cm temperature (Celsius)}
   \item{t1m}{1m temperature (Celsius)}
   \item{wr}{Wind run (kilometres)}
   \item{state}{State name (postal code abbreviation)}
   \item{lat}{Latitude (decimal degrees)}
   \item{lon}{Longitude (decimal degrees)}
}
}
\description{
Fetch the BOM agricultural bulletin information
}
\details{
Allowed state and territory postal codes, only one state per request
or all using \code{AUS}.
 \describe{
   \item{NSW}{New South Wales}
   \item{NT}{Northern Territory}
   \item{QLD}{Queensland}
   \item{SA}{South Australia}
   \item{TAS}{Tasmania}
   \item{VIC}{Victoria}
   \item{WA}{Western Australia}
   \item{AUS}{Australia, returns bulletin for all states}
 }
}
\examples{
\dontrun{
ag_bulletin <- get_bulletin(state = "QLD")
}

}
\references{
Australian Bureau of Meteorology (BOM) Weather Data Services Agriculture Bulletins
\url{http://www.bom.gov.au/catalogue/observations/about-agricultural.shtml}

Australian Bureau of Meteorology (BOM) Weather Data Services Observation of Rainfall
\url{http://www.bom.gov.au/climate/how/observations/rain-measure.shtml}
}
\author{
Adam H Sparks, \email{adamhsparks@gmail.com}
}
