% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_current_weather.R
\name{get_current_weather}
\alias{get_current_weather}
\title{Current Weather Observations of a BoM Station}
\usage{
get_current_weather(station_name, strict = FALSE, latlon = NULL,
  raw = FALSE, emit_latlon_msg = TRUE, as.data.table = FALSE)
}
\arguments{
\item{station_name}{The name of the weather station. Fuzzy string matching
via \code{base::agrep} is done.}

\item{strict}{(logical) If \code{TRUE}, \code{station_name} must match the station name exactly,
except that \code{station_name} need not be uppercase. Note this may be different to
\code{full_name} in the response. See \strong{Details}.}

\item{latlon}{A length-2 numeric vector giving the decimal
latitude and longitude (in that order), \emph{e.g.} \code{latlon = c(-34, 151)} for Sydney.
When given instead of \code{station_name}, the nearest station (in this package) is used, with a
message indicating the nearest such station. (See also
 \code{\link{sweep_for_stations}}.) Ignored if used in combination with
\code{station_name}, with a warning.}

\item{raw}{Logical. Do not convert the columns \code{data.table} to the
appropriate classes. (\code{FALSE} by default.)}

\item{emit_latlon_msg}{Logical. If \code{TRUE} (the default), and
\code{latlon} is selected, a message is emitted before the table is returned
indicating which station was actually used (i.e. which station was found to
be nearest to the given coordinate).}

\item{as.data.table}{Return result as a \code{data.table}.}
}
\value{
full details of fields and units returned, see Appendix 1 in the
\emph{bomrang} vignette, use \code{vignette("bomrang", package = "bomrang")}
to view.
}
\description{
Current Weather Observations of a BoM Station
}
\details{
Station names are not consistently named within the Bureau, so
the response may contain a different \code{full_name} to the one
matched, even if \code{strict = TRUE}. For example,
\code{get_current_weather("CASTLEMAINE PRISON")[["full_name"]][1]}
is \code{Castlemaine}, not \code{Castlemaine Prison}.

Note that the column \code{local_date_time_full} is set to a
\code{POSIXct} object in the local time of the \strong{user}.
For more details see the vignette "Current Weather Fields":
\code{vignette("Current Weather Fields", package = "bomrang")}
for a complete list of fields and units.
}
\examples{
\dontrun{
  # warning
  Melbourne_weather <- get_current_weather("Melbourne")

  # no warning
  Melbourne_weather <- get_current_weather("Melbourne (Olympic Park)")

  # Get weather by latitude and longitude:
  get_current_weather(latlon = c(-34, 151))
}
}
\references{
Weather data observations are retrieved from:
Australian Bureau of Meteorology (BoM) Weather Data Services,
Observations - individual stations:
\url{http://www.bom.gov.au/catalogue/data-feeds.shtml}

Station location and other metadata are sourced from the Australian Bureau of
Meteorology (BoM) webpage, Bureau of Meteorology Site Numbers:
\url{http://www.bom.gov.au/climate/cdo/about/site-num.shtml}
}
\author{
Hugh Parsonage, \email{hugh.parsonage@gmail.com}
}
