% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_BOM_forecast.R
\name{get_forecast}
\alias{get_forecast}
\title{Get BOM Forecast}
\usage{
get_forecast(state)
}
\arguments{
\item{state}{Australian state or territory, either as full name or abbreviation}
}
\value{
Data frame of a Australia BOM forecast for max temperature, min temperature
and corresponding locations with lat/lon values for the next six days.
}
\description{
Fetch the BOM forecast and create a data frame object that can be used for
interpolating.
}
\examples{
\dontrun{
BOM_forecast <- get_forecast(state = "Qld")
}

}
