% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_forecast.R
\name{get_forecast}
\alias{get_forecast}
\title{Get BOM Forecast}
\usage{
get_forecast(state = NULL)
}
\arguments{
\item{state}{Australian state or territory as postal code, see details for
instruction.}
}
\value{
Data frame of a Australia BOM daily forecast in a data frame with the
following fields.

\describe{
   \item{aac}{AMOC Area Code, e.g. WA_MW008, a unique identifier for each location}
   \item{date}{Date (YYYY-MM-DD)}
   \item{max_temp}{Maximum forecasted temperature (degrees Celsius)}
   \item{min_temp}{Minimum forecasted temperature (degrees Celsius)}
   \item{lower_prcp_limit}{Lower forecasted precipitation limit (millimetres)}
   \item{upper_prcp_limit}{Upper forecasted precipitation limit (millimetres)}
   \item{precis}{Précis forecast (a short summary, less than 30 characters)}
   \item{prob_prcp}{Probability of precipitation (percent)}
   \item{location}{Named location for forecast}
   \item{state}{State name  (postal code abbreviation)}
   \item{lon}{Longitude of named location (decimal Degrees)}
   \item{lat}{Latitude of named location (decimal Degrees)}
   \item{elev}{Elevation of named location (metres)}
}
}
\description{
Fetch the BOM précis forecast and return a tidy data frame of the daily
forecast
}
\details{
Allowed state and territory postal codes, only one state per request
or all using \code{AUS}.
 \describe{
   \item{ACT}{Australian Capital Territory}
   \item{NSW}{New South Wales}
   \item{NT}{Northern Territory}
   \item{QLD}{Queensland}
   \item{SA}{South Australia}
   \item{TAS}{Tasmania}
   \item{VIC}{Victoria}
   \item{WA}{Western Australia}
   \item{AUS}{Australia, returns forecast for all states}
 }
}
\examples{
\dontrun{
BOM_forecast <- get_forecast(state = "QLD")
}

}
\references{
Australian Bureau of Meteorology (BOM) Weather Data Services
\url{http://www.bom.gov.au/catalogue/data-feeds.shtml}
}
\author{
Adam H Sparks, \email{adamhsparks@gmail.com} and Keith Pembleton \email{keith.pembleton@usq.edu.au}
}
