% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_historical.R
\name{get_historical}
\alias{get_historical}
\title{Obtain Historical BOM Data}
\usage{
get_historical(stationid = NULL, latlon = NULL, type = c("rain", "min",
  "max", "solar"))
}
\arguments{
\item{stationid}{BOM station ID. See Details.}

\item{latlon}{Length-2 numeric vector of Latitude/Longitude. See Details.}

\item{type}{Measurement type, either daily "rain", "min" (temp), "max"
(temp), or "solar" (exposure). Partial matching is performed.}
}
\value{
a complete \link{data.frame} of historical observations for the chosen
station.

Either \code{stationid} or \code{latlon} must be provided, but if both are, then
\code{stationid} will be used as it is more reliable.

In some cases data is available back to the 1800s, so tens of thousands of
daily records will be returned. Other stations will be newer and will
return fewer observations.
}
\description{
Retrieves daily observations for a given station.
}
\examples{
\dontrun{get_historical(stationid = "023000", type = "max") ## ~48,000+ daily records}
get_historical(latlon = c(-35.2809, 149.1300), type = "min") ## 3,500+ daily records
}
\author{
Jonathan Carroll, \email{rpkg@jcarroll.com.au}
}
