% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_subdaily_weather.R
\name{get_subdaily_weather}
\alias{get_subdaily_weather}
\title{Obtain historical BOM data in sub-daily time scales}
\usage{
get_subdaily_weather(
  stationid = NULL,
  name = NULL,
  years = NULL,
  hourly = TRUE,
  ...
)
}
\arguments{
\item{stationid}{\acronym{BOM} station \sQuote{ID}. See Details.}

\item{years}{Year(s) of weather data to download. Entered as integers,
\emph{e.g.} \code{2001:2002}. Defaults to all available years for the
specified station if left unspecified.}

\item{hourly}{If \code{TRUE} this forces the values to start on
the hour and returns hourly values for the entire period requested rather
than sub-hourly values. Defaults to \code{TRUE}.}

\item{...}{Other parameters as passed along to
\code{\link[stationaRy]{get_met_data}}.}

\item{station_name}{Official \acronym{BOM} station name. See Details.}
}
\value{
A \code{bomrang_tbl} object (extension of a
\code{\link[base]{data.frame}}) of historical sub-daily (hourly or
less)time-scale observations for the selected station.
}
\description{
Retrieves hourly or smaller, if available, weather observations for a given
station.
}
\details{
A value for \var{stationid} or \var{name} must be provided. If
you are uncertain, you may use \code{\link{sweep_for_stations}} to identify
possible candidate stations to query. See
\code{vignette("bomrang", package = "bomrang")} for an example of this use.

Only stations falling within BOM's network are returned. If you require
stations that are outside of BOM's network, please use
\CRANpkg{stationaRy} directly.
}
\examples{
\donttest{
# Get hourly weather data, both `stationid` and `name` refer to the same
# station, CHARLTON, in Victoria
get_subdaily_weather(stationid = "080128")
get_subdaily_weather(name = "CHARLTON")
}

}
\seealso{
\link{get_current_weather} \link{get_historical_weather}
}
