% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{filter}
\alias{reexports}
\alias{select}
\alias{reexports}
\alias{mutate}
\alias{reexports}
\alias{rename}
\alias{reexports}
\alias{arrange}
\alias{reexports}
\alias{group_by}
\alias{reexports}
\alias{slice}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{dplyr}{\code{\link[dplyr]{filter}}, \code{\link[dplyr]{select}}, \code{\link[dplyr]{mutate}}, \code{\link[dplyr]{rename}}, \code{\link[dplyr]{arrange}}, \code{\link[dplyr]{group_by}}, \code{\link[dplyr]{slice}}}
}}

