% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_satellite_imagery.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{plot}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{terra}{\code{\link[terra]{plot}}}
}}

