% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{stations_site_list}
\alias{stations_site_list}
\title{stations_site_list}
\format{An object of class \code{data.frame} with 7439 rows and 11 columns.}
\source{
\url{ftp://ftp.bom.gov.au/anon2/home/ncc/metadata/sitelists/stations.zip}
}
\usage{
stations_site_list
}
\description{
A database of station names, locations and other metadata used for the ag
bulletin
}
\details{
\describe{
   \item{site}{Unique BoM identifier for each station}
   \item{dist}{BoM rainfall district}
   \item{name}{BoM station name}
   \item{start}{Year data collection starts}
   \item{end}{Year data collection ends (will always be current)}
   \item{state}{State name (postal code abbreviation)}
   \item{lat}{Latitude (decimal degrees)}
   \item{lon}{Longitude (decimal degrees)}
   \item{elev_m}{Station elevation (metres)}
   \item{bar_ht}{Bar height (metres)}
   \item{WMO}{World Meteorological Organization number (unique ID used worldwide)}
   }
}
\keyword{datasets}
