% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_forecast_towns.R
\name{update_forecast_towns}
\alias{update_forecast_towns}
\title{Update internal database with latest BOM forecast towns}
\usage{
update_forecast_towns()
}
\value{
Updated database of \acronym{BOM} précis forecast towns
}
\description{
Download the latest select forecast towns from the \acronym{BOM} server and
update internal database of précis forecast town names and \acronym{AAC}
codes used by \code{\link{get_precis_forecast}}.  There is no need to use
this unless you know that a forecast town exists in a more current version of
the \acronym{BOM} précis forecast town name database that is not available in
the database distributed with \pkg{bomrang}.  In fact, for reproducibility
purposes, users are discouraged from using this function.
}
\examples{
\dontrun{
update_forecast_towns()
}
}
\references{
Data are sourced from: Australian Bureau of Meteorology (\acronym{BOM})
webpage, "Weather Data Services",
\url{http://www.bom.gov.au/catalogue/data-feeds.shtml}
}
\author{
Adam H. Sparks, \email{adamhsparks@gmail.com}
}
