% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_station_locations.R
\name{update_station_locations}
\alias{update_station_locations}
\title{Update Internal Database With Latest BOM Station Locations and Metadata}
\usage{
update_station_locations()
}
\value{
Updated internal database of BOM station locations
}
\description{
Download the latest station locations and metadata and update bomrang's
internal database.  There is no need to use this unless you know that a
station exists in BOM's database that is not available in the database
distributed with \code{\link{bomrang}}.
}
\examples{
\dontrun{
update_station_locations()
}
}
\references{
Australian Bureau of Meteorology (BOM) Site Numbers
\url{http://www.bom.gov.au/climate/cdo/about/site-num.shtml}
}
\author{
Adam H Sparks, \email{adamhsparks@gmail.com}
}
